+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### NEXT
 
+* Re-organised logic for handling analytics
+  * `staticId` has been renamed to `id`
 * Included timestamp details in analytic results
 * The support for periodical (once a day, at midnight) re-computation of _moving time window_ analytic results have been added.
 * _Time window_ type values have been updated. In particular: